requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sentence-transformers==5.0.0
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
transformers==4.54.1
numpy==2.3.2
scikit-learn==1.7.1
scipy==1.16.1
python-dotenv==1.0.1
requests==2.32.4 